+++ src/lightly_train/_commands/common_helpers.py
@@ -8,41 +8,37 @@
 from __future__ import annotations
 
 import contextlib
+import hashlib
 import json
 import logging
 import os
-import platform
+import tempfile
+import time
 import warnings
 from enum import Enum
 from pathlib import Path
-from tempfile import NamedTemporaryFile
-from typing import Any, Generator, Iterable, Literal
+from typing import Any, Generator, Iterable, Literal, Sequence
 
 import torch
 from pytorch_lightning.accelerators.accelerator import Accelerator
 from pytorch_lightning.accelerators.cpu import CPUAccelerator
 from pytorch_lightning.accelerators.cuda import CUDAAccelerator
 from pytorch_lightning.accelerators.mps import MPSAccelerator
-from pytorch_lightning.plugins.environments import (  # type: ignore[attr-defined]
-    SLURMEnvironment,
-)
 from pytorch_lightning.strategies.strategy import Strategy
 from torch.nn import Module
 from torch.utils.data import Dataset
 
-from lightly_train._commands import common_helpers
+from lightly_train._commands import _lightning_rank_zero
 from lightly_train._data import image_dataset
 from lightly_train._data._serialize import memory_mapped_sequence
 from lightly_train._data._serialize.memory_mapped_sequence import MemoryMappedSequence
 from lightly_train._data.image_dataset import ImageDataset
 from lightly_train._embedding.embedding_format import EmbeddingFormat
+from lightly_train._env import Env
 from lightly_train._models import package_helpers
 from lightly_train.types import DatasetItem, PathLike, Transform
 
 logger = logging.getLogger(__name__)
-
-
-LIGHTLY_TRAIN_MASK_DIR = os.environ.get("LIGHTLY_TRAIN_MASK_DIR", None)
 
 
 def get_checkpoint_path(checkpoint: PathLike) -> Path:
@@ -89,26 +85,47 @@ def get_accelerator(
         return CPUAccelerator()
 
 
-def _get_rank() -> int | None:
-    """Get the rank of the current process.
+get_global_rank = _lightning_rank_zero.get_global_rank
 
-    Copied from https://github.com/Lightning-AI/pytorch-lightning/blob/06a8d5bf33faf0a4f9a24207ae77b439354350af/src/lightning/fabric/utilities/rank_zero.py#L39-L49
-    """
-    # SLURM_PROCID can be set even if SLURM is not managing the multiprocessing,
-    # therefore LOCAL_RANK needs to be checked first
-    rank_keys = ("RANK", "LOCAL_RANK", "SLURM_PROCID", "JSM_NAMESPACE_RANK")
+
+def get_local_rank() -> int | None:
+    """Get the local rank of the current process."""
+    rank_keys = ("LOCAL_RANK", "SLURM_LOCALID", "JSM_NAMESPACE_LOCAL_RANK")
     for key in rank_keys:
         rank = os.environ.get(key)
         if rank is not None:
             return int(rank)
-    # None to differentiate whether an environment variable was set at all
     return None
 
 
-def is_rank_zero() -> bool:
-    """Check if the current process is running on the first device."""
-    local_rank = _get_rank()
+def get_node_rank() -> int | None:
+    """Get the node rank of the current process."""
+    rank_keys = ("NODE_RANK", "GROUP_RANK", "SLURM_NODEID")
+    for key in rank_keys:
+        rank = os.environ.get(key)
+        if rank is not None:
+            return int(rank)
+    return None
+
+
+def is_global_rank_zero() -> bool:
+    """Check if the current process is running on the global rank zero."""
+    global_rank = get_global_rank()
+    # Check node rank because process might be assigned to a node but not yet
+    # a global rank.
+    return global_rank == 0 or (global_rank is None and is_node_rank_zero())
+
+
+def is_local_rank_zero() -> bool:
+    """Check if the current process is running on the local rank zero."""
+    local_rank = get_local_rank()
     return local_rank == 0 or local_rank is None
+
+
+def is_node_rank_zero() -> bool:
+    """Check if the current process is running on the node rank zero."""
+    node_rank = get_node_rank()
+    return node_rank == 0 or node_rank is None
 
 
 def get_out_dir(out: PathLike, resume: bool, overwrite: bool) -> Path:
@@ -120,7 +137,7 @@ def get_out_dir(out: PathLike, resume: bool, overwrite: bool) -> Path:
 
         dir_not_empty = any(out_dir.iterdir())
 
-        if dir_not_empty and (not (resume or overwrite)) and is_rank_zero():
+        if dir_not_empty and (not (resume or overwrite)) and is_global_rank_zero():
             raise ValueError(
                 f"Output '{out_dir}' is not empty! Set overwrite=True to overwrite the "
                 "directory or resume=True to resume training."
@@ -129,10 +146,109 @@ def get_out_dir(out: PathLike, resume: bool, overwrite: bool) -> Path:
     return out_dir
 
 
+def get_tmp_dir() -> Path:
+    """Get the temporary directory for Lightly Train."""
+    return Path(tempfile.gettempdir()) / "lightly-train"
+
+
+def get_data_tmp_dir() -> Path:
+    """Get the temporary directory for Lightly Train data."""
+    return get_tmp_dir() / "data"
+
+
+def get_verify_out_tmp_dir() -> Path:
+    """Get the temporary directory for Lightly Train verify out."""
+    return get_tmp_dir() / "verify-out"
+
+
+def get_sha256(value: Any) -> str:
+    """Get the SHA256 hash of a value."""
+    return hashlib.sha256(str(value).encode()).hexdigest()
+
+
+@contextlib.contextmanager
+def verify_out_dir_equal_on_all_local_ranks(out: Path) -> Generator[None, None, None]:
+    """Verify that the out path is the same on all local ranks.
+
+    This is important for distributed training, as the out path is used as
+    a deterministic value that must be consistent across all local ranks.
+
+    A common case where this can fail is when the out path contains a timestamp
+    in the path that is generated inside the training script. For example with:
+    >>> out = f"out/{datetime.now().strftime('%Y-%m-%d_%H-%M-%S')}"
+    This will result in different paths for each rank as the timestamp is generated
+    on each rank separately.
+    """
+    out_dir = Path(out).resolve()
+    # Add the node rank to the filename. This makes sure that each node verifies
+    # its out directory separately, even if the nodes are using a shared filesystem.
+    out_tmp = get_verify_out_tmp_dir() / get_sha256(f"{out_dir}-{get_node_rank() or 0}")
+    logger.debug(f"Creating temporary file '{out_tmp}' to verify out path.")
+
+    try:
+        if is_local_rank_zero():
+            _unlink_and_ignore(out_tmp)
+            out_tmp.parent.mkdir(parents=True, exist_ok=True)
+            out_tmp.touch()
+            # Write the out directory to the temporary file for debugging.
+            out_tmp.write_text(str(out_dir))
+            yield
+        else:
+            # Wait for rank zero to create the temporary file.
+            timeout_sec = Env.LIGHTLY_TRAIN_VERIFY_OUT_DIR_TIMEOUT_SEC.value
+            start_time_sec = time.time()
+            while not out_tmp.exists():
+                if timeout_sec >= 0 and time.time() - start_time_sec > timeout_sec:
+                    raise RuntimeError(
+                        f"Rank {get_global_rank()}: Timeout after {timeout_sec} seconds "
+                        "while verifying that all ranks (processes) have the same 'out' path. "
+                        "This means that the 'out' path is not set to the same path on all ranks. "
+                        "If the path to your 'out' path contains any timestamps make sure that "
+                        "they are provided from OUTSIDE of the training script. Either via the "
+                        "command line or an environment variable. Timestamps created inside a Python "
+                        "script, for example with 'datetime.now()' or 'time.time()', will result "
+                        "in different values for each rank and must not be used. "
+                        f"The timeout can be configured with the {Env.LIGHTLY_TRAIN_VERIFY_OUT_DIR_TIMEOUT_SEC.name} "
+                        "environment variable. Setting it to -1 disables the timeout. "
+                    )
+                time.sleep(0.1)
+            yield
+    finally:
+        _unlink_and_ignore(out_tmp)
+
+
 def pretty_format_args(args: dict[str, Any], indent: int = 4) -> str:
     args = sanitize_config_dict(args)
 
     return json.dumps(args, indent=indent, sort_keys=True)
+
+
+def remove_excessive_args(
+    args: dict[str, Any], limit_keys: set[str] | None = None, num_elems: int = 5
+) -> dict[str, Any]:
+    """Limit the number of elements in sequences of a dict to a certain number. This is
+    strictly for logging purposes. Does not work with nested structures.
+
+    Args:
+        args: The dictionary to limit.
+        limit_keys: The keys to limit. If None, all keys are limited.
+        num_elems: Number of elements to keep in the sequence.
+
+    Returns:
+        The dictionary with sequences limited to a certain number of elements.
+    """
+    if limit_keys is None:
+        limit_keys = set(args.keys())
+    for key in limit_keys:
+        if (
+            isinstance(args[key], Sequence)
+            and not isinstance(args[key], str)
+            and len(args[key]) > num_elems
+        ):
+            args[key] = type(args[key])(
+                (*args[key][: num_elems - 2], "...", args[key][-1])
+            )
+    return args
 
 
 def sanitize_config_dict(args: dict[str, Any]) -> dict[str, Any]:
@@ -216,8 +332,10 @@ class ModelFormat(Enum):
         raise ValueError(f"{value} is not a valid {cls.__name__}")
 
 
-def export_model(model: Module, format: ModelFormat, out: Path) -> None:
-    if not is_rank_zero():
+def export_model(
+    model: Module, format: ModelFormat, out: Path, log_example: bool = True
+) -> None:
+    if not is_global_rank_zero():
         return
     logger.debug(f"Exporting model to '{out}' in format '{format}'.")
     out.parent.mkdir(parents=True, exist_ok=True)
@@ -227,46 +345,42 @@ def export_model(model: Module, format: ModelFormat, out: Path) -> None:
         torch.save(model.state_dict(), out)
     elif format == ModelFormat.PACKAGE_DEFAULT:
         package = package_helpers.get_package_from_model(model=model)
-        package.export_model(model=model, out=out)
+        package.export_model(model=model, out=out, log_example=log_example)
     else:
         raise ValueError(f"Invalid format: '{format.value}' is not supported ")
 
 
 @contextlib.contextmanager
-def get_dataset_temp_mmap_path() -> Generator[Path, Any, Any]:
+def get_dataset_temp_mmap_path(out: Path) -> Generator[Path, Any, Any]:
     """Generate file in temporary directory to be used for memory-mapping the dataset.
 
-    In case the Lightning Trainer is used to spawn the processes, rank 0 will run first
-    and create the memory-mapped file and share this via the env variable
-    LIGHTLY_TRAIN_DATASET_MMAP_PATH. This allows all processes to use the same file.
+    Creates a unique filename for the memory-mapped file based on the out path.
+    We use the out path as a deterministic value that is consistent across all ranks
+    on the same node.
 
-    In case training runs on SLURM with the `srun` command, the control over spawning
-    the processes is given to SLURM and `srun` immediately runs the script `--ntasks-per-node`
-    many times in parallel. Therefore a common location of the temporary file can't be
-    communicated between the processes and every process has to create its own memory-mapped
-    file.
-    On Windows, each process creates its own temporary file due to file handling restrictions.
+    We need a determinstic value from "outside" at this point in the code as the
+    code might already be running on multiple processes depending on how it was
+    started. We cannot create a new filename based on a random value as this would
+    create a different filename for each process. Creating the filename on global
+    rank zero and sharing it across all ranks is also complicated here as
+    torch.distributed is not necessarily initialized yet and there are many forms
+    of parallelism to handle (fork/spawn, torch.distributed, SLURM, etc.).
+
+    The filename is different on each node. This is necessary to avoid multiple
+    processes writing to the same file in case the nodes use a shared filesystem.
     """
-    if platform.system() == "Windows" or SLURMEnvironment.detect():
-        # On Windows or SLURM, every rank creates its own temporary file
-        with NamedTemporaryFile(delete=True) as mmap_file:
-            mmap_filepath = Path(mmap_file.name)
-            # Close the file here to prevent failures on Windows, as it will be accessed
-            # by another process.
-            mmap_file.close()
-            yield mmap_filepath
-    else:
-        # Unix-like systems can share the same file between processes
-        LIGHTLY_TRAIN_DATASET_MMAP_PATH = "LIGHTLY_TRAIN_DATASET_MMAP_PATH"
-        if common_helpers.is_rank_zero():
-            with NamedTemporaryFile(delete=True) as mmap_file:
-                mmap_filepath = Path(mmap_file.name)
-                os.environ[LIGHTLY_TRAIN_DATASET_MMAP_PATH] = str(mmap_filepath)
-                yield mmap_filepath
-        else:
-            # Make sure that all ranks use the same memory-mapped file.
-            mmap_filepath = Path(os.environ[LIGHTLY_TRAIN_DATASET_MMAP_PATH])
-            yield mmap_filepath
+    out_hash = get_sha256(f"{out}-{get_node_rank() or 0}")
+    mmap_filepath = (get_data_tmp_dir() / out_hash).with_suffix(".mmap")
+    mmap_filepath.parent.mkdir(parents=True, exist_ok=True)
+    try:
+        # Delete the file if it already exists from a previous run.
+        if is_local_rank_zero():
+            _unlink_and_ignore(mmap_filepath)
+
+        yield mmap_filepath
+    finally:
+        if is_local_rank_zero():
+            _unlink_and_ignore(mmap_filepath)
 
 
 def get_dataset_mmap_filenames(
@@ -275,34 +389,49 @@ def get_dataset_mmap_filenames(
 ) -> MemoryMappedSequence[str]:
     """Returns memory-mapped filenames shared across all ranks.
 
-    In case training runs on SLURM with the `srun` command, the control over spawning
-    the processes is given to SLURM and `srun` immediately runs the script `--ntasks-per-node`
-    many times in parallel. Therefore a common location of the temporary file can't be
-    communicated between the processes and every process has to create its own memory-
-    mapped file.
+    Filenames are written to mmap_filepath by rank zero and read by all ranks.
     """
-    if platform.system() == "Windows" or SLURMEnvironment.detect():
-        # On Windows or SLURM, every rank creates its own temporary file
-        return memory_mapped_sequence.memory_mapped_sequence_from_filenames(
-            filenames=filenames,
-            mmap_filepath=mmap_filepath,
-        )
-    else:
-        if common_helpers.is_rank_zero():
-            # Save filenames to memory mapped file and return them.
-            return memory_mapped_sequence.memory_mapped_sequence_from_filenames(
+    tmp_path = mmap_filepath.with_suffix(".temp")
+    try:
+        if is_local_rank_zero():
+            # Save filenames to temporary file. Create the final file only once rank zero has
+            # finished writing all the filenames.
+            memory_mapped_sequence.write_filenames_to_file(
                 filenames=filenames,
-                mmap_filepath=mmap_filepath,
+                mmap_filepath=tmp_path,
             )
+            # Rename the temporary file to mmap_filepath.
+            tmp_path.replace(mmap_filepath.resolve())
         else:
-            # Return memory-mapped filenames from file.
-            return memory_mapped_sequence.memory_mapped_sequence_from_file(
-                mmap_filepath=mmap_filepath
-            )
+            # Wait for rank zero to finish writing the filenames.
+            timeout_sec = Env.LIGHTLY_TRAIN_MMAP_TIMEOUT_SEC.value
+            start_time_sec = time.time()
+            while not mmap_filepath.exists():
+                if tmp_path.exists():
+                    # Reset timeout if the temporary file exists. This means that rank zero
+                    # is still writing the filenames.
+                    start_time_sec = time.time()
+
+                if timeout_sec >= 0 and time.time() - start_time_sec > timeout_sec:
+                    raise RuntimeError(
+                        f"Rank {get_global_rank()}: Timeout after {timeout_sec} seconds "
+                        f"while waiting for the memory-mapped file '{mmap_filepath}' to be created. "
+                        "Please contact Lightly support if this happens. This is most likely a bug. "
+                        f"You can increase the timeout with the {Env.LIGHTLY_TRAIN_MMAP_TIMEOUT_SEC.name} "
+                        "environment variable. Setting it to -1 disables the timeout. "
+                    )
+                time.sleep(0.2)
+    finally:
+        _unlink_and_ignore(tmp_path)
+
+    # Return memory-mapped filenames from file.
+    return memory_mapped_sequence.memory_mapped_sequence_from_file(
+        mmap_filepath=mmap_filepath
+    )
 
 
 def get_dataset(
-    data: PathLike | Dataset[DatasetItem],
+    data: PathLike | Sequence[PathLike] | Dataset[DatasetItem],
     transform: Transform,
     mmap_filepath: Path | None,
 ) -> Dataset[DatasetItem]:
@@ -310,28 +439,70 @@ def get_dataset(
         logger.debug("Using provided dataset.")
         return data
 
-    data = Path(data).resolve()
-    logger.debug(f"Making sure data directory '{data}' exists and is not empty.")
-    if not data.exists():
-        raise ValueError(f"Data directory '{data}' does not exist!")
-    elif not data.is_dir():
-        raise ValueError(f"Data path '{data}' is not a directory!")
-    elif data.is_dir() and not any(data.iterdir()):
-        raise ValueError(f"Data directory '{data}' is empty!")
     if mmap_filepath is None:
         raise ValueError("Memory-mapped file path must be provided.")
 
-    logger.info(f"Initializing dataset from '{data}'.")
-    # NOTE(Guarin, 01/25): The bottleneck for dataset initialization is filename
-    # listing and not the memory mapping. Listing the train set from ImageNet takes
-    # about 30 seconds. This is mostly because os.walk is not parallelized.
-    filenames = image_dataset.list_image_filenames(image_dir=data)
-    return ImageDataset(
-        image_dir=data,
-        image_filenames=get_dataset_mmap_filenames(
-            filenames=filenames,
-            mmap_filepath=mmap_filepath,
-        ),
-        transform=transform,
-        mask_dir=Path(LIGHTLY_TRAIN_MASK_DIR) if LIGHTLY_TRAIN_MASK_DIR else None,
-    )
+    mask_dir = Env.LIGHTLY_TRAIN_MASK_DIR.value
+
+    if isinstance(data, (str, Path)):
+        data = Path(data).resolve()
+        if not data.exists():
+            raise ValueError(f"Data directory '{data}' does not exist!")
+        elif not data.is_dir():
+            raise ValueError(f"Data path '{data}' is not a directory!")
+        elif data.is_dir() and not any(data.iterdir()):
+            raise ValueError(f"Data directory '{data}' is empty!")
+        # Use relative paths as filenames when a single directory or file is provided to
+        # reduce the file size.
+        # NOTE(Guarin, 01/25): The bottleneck for dataset initialization is filename
+        # listing and not the memory mapping. Listing the train set from ImageNet takes
+        # about 30 seconds. This is mostly because os.walk is not parallelized.
+        filenames = image_dataset.list_image_filenames(image_dir=data)
+        return ImageDataset(
+            image_dir=data,
+            image_filenames=get_dataset_mmap_filenames(
+                filenames=filenames,
+                mmap_filepath=mmap_filepath,
+            ),
+            transform=transform,
+            mask_dir=Path(mask_dir) if mask_dir is not None else None,
+        )
+
+    elif isinstance(data, Sequence):
+        _data: Sequence[Path] = [Path(d).resolve() for d in data]
+        if mask_dir is not None:
+            raise ValueError(
+                "Mask directory is not supported when multiple directories or files "
+                "are provided."
+            )
+
+        for d in _data:
+            if not d.exists():
+                raise ValueError(f"Data directory or file '{d}' does not exist!")
+            elif d.is_dir() and not any(d.iterdir()):
+                raise ValueError(f"Data directory '{d}' is empty!")
+        files = image_dataset.list_image_files(imgs_and_dirs=_data)
+        filenames = image_dataset.list_image_filenames(files=files)
+        return ImageDataset(
+            image_dir=None,
+            image_filenames=get_dataset_mmap_filenames(
+                filenames=filenames,
+                mmap_filepath=mmap_filepath,
+            ),
+            transform=transform,
+        )
+    else:
+        raise ValueError(
+            "Data must be a directory, a list of directories or files, or a dataset."
+        )
+
+
+def _unlink_and_ignore(path: Path) -> None:
+    """Unlink a file and ignore the error if it fails.
+
+    Errors can happen if we do not have permission to access the file.
+    """
+    try:
+        path.unlink(missing_ok=True)
+    except OSError:
+        pass