+++ src/lightly_train/_models/timm/timm_package.py
@@ -16,9 +16,9 @@ import torch
 from torch.nn import Module
 
 from lightly_train._models import package_helpers
-from lightly_train._models.feature_extractor import FeatureExtractor
+from lightly_train._models.model_wrapper import ModelWrapper
 from lightly_train._models.package import Package
-from lightly_train._models.timm.timm import TIMMFeatureExtractor
+from lightly_train._models.timm.timm import TIMMModelWrapper
 
 logger = logging.getLogger(__name__)
 
@@ -67,35 +67,36 @@ class TIMMPackage(Package):
         return model
 
     @classmethod
-    def get_feature_extractor(cls, model: Module) -> FeatureExtractor:
-        return TIMMFeatureExtractor(model)
+    def get_model_wrapper(cls, model: Module) -> ModelWrapper:
+        return TIMMModelWrapper(model)
 
     @classmethod
-    def export_model(cls, model: Module, out: Path) -> None:
+    def export_model(cls, model: Module, out: Path, log_example: bool = True) -> None:
         torch.save(model.state_dict(), out)
 
-        model_name = model.pretrained_cfg.get("architecture", None)
-        if not model_name:
-            logger.warning(
-                "Usage example can not be constructed since the model name is unknown."
-            )
-            return
+        if log_example:
+            model_name = model.pretrained_cfg.get("architecture", None)
+            if not model_name:
+                logger.warning(
+                    "Usage example can not be constructed since the model name is unknown."
+                )
+                return
 
-        log_message_code = [
-            "import timm",
-            "",
-            "# Load the pretrained model",
-            "model = timm.create_model(",
-            f"    model_name='{model_name}',",
-            f"    checkpoint_path='{out}',",
-            ")",
-            "",
-            "# Finetune or evaluate the model",
-            "...",
-        ]
-        logger.info(
-            package_helpers.format_log_msg_model_usage_example(log_message_code)
-        )
+            log_message_code = [
+                "import timm",
+                "",
+                "# Load the pretrained model",
+                "model = timm.create_model(",
+                f"    model_name='{model_name}',",
+                f"    checkpoint_path='{out}',",
+                ")",
+                "",
+                "# Finetune or evaluate the model",
+                "...",
+            ]
+            logger.info(
+                package_helpers.format_log_msg_model_usage_example(log_message_code)
+            )
 
 
 # Create singleton instance of the package. The singleton should be used whenever