+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Support lists of files and folders as input to the `embed` and `train` commands.
+- Faster dataset initialization with SLURM and Windows.
+- Model is exported after every n-th epoch instead of only at the end. Default n=1
+
 ### Changed
 
 - Distillation uses the official DINOv2 implementation for the teacher model.